refactor: compute highlighted ranges from esquery matched nodes
